+++ packages/matter-node-shell.js/src/shell/cmd_config.ts
@@ -301,7 +301,7 @@ function doWifiCredentials(
                 `Current Wifi-Credentials: SSID="${theNode.Store.get<string>(
                     "WiFiSsid",
                     "-",
-                )}, Password="${Logger.maskString(theNode.Store.get<string>("WiFiPassword", ""))}"`,
+                )}", Password="${Logger.maskString(theNode.Store.get<string>("WiFiPassword", ""))}"`,
             );
             break;
         case "set":
@@ -315,7 +315,7 @@ function doWifiCredentials(
                 `New Wifi-Credentials: SSID="${theNode.Store.get<string>(
                     "WiFiSsid",
                     "-",
-                )}, Password="${Logger.maskString(theNode.Store.get<string>("WiFiPassword"))}"`,
+                )}", Password="${Logger.maskString(theNode.Store.get<string>("WiFiPassword"))}"`,
             );
             break;
         case "delete":
@@ -342,7 +342,7 @@ function doThreadCredentials(
                 `Current Thread network credentials: name="${theNode.Store.get<string>(
                     "ThreadName",
                     "-",
-                )}, Operational-Dataset="${Logger.maskString(
+                )}", Operational-Dataset="${Logger.maskString(
                     theNode.Store.get<string>("ThreadOperationalDataset", ""),
                 )}"`,
             );
@@ -358,7 +358,7 @@ function doThreadCredentials(
                 `New Wifi-Credentials: SSID="${theNode.Store.get<string>(
                     "ThreadName",
                     "-",
-                )}, OperationalDataset="${Logger.maskString(theNode.Store.get<string>("ThreadOperationalDataset"))}"`,
+                )}", OperationalDataset="${Logger.maskString(theNode.Store.get<string>("ThreadOperationalDataset"))}"`,
             );
             break;
         case "delete":