+++ packages/matter.js/src/protocol/interaction/InteractionClient.ts
@@ -459,7 +459,7 @@ export class InteractionClient {
             const writeRequests = attributes.map(
                 ({ endpointId, clusterId, attribute: { id, schema }, value, dataVersion }) => ({
                     path: { endpointId, clusterId, attributeId: id },
-                    data: schema.encodeTlv(value),
+                    data: schema.encodeTlv(value, true),
                     dataVersion,
                 }),
             );
@@ -864,6 +864,9 @@ export class InteractionClient {
                     commandId: requestId,
                 })} with ${Logger.toJSON(request)}`,
             );
+
+            requestSchema.validate(request);
+
             const commandFields = requestSchema.encodeTlv(request);
 
             if (timedRequest) {