+++ packages/matter.js/src/tlv/TlvObject.ts
@@ -5,7 +5,9 @@
  */
 
 import { UnexpectedDataError } from "../common/MatterError.js";
+import { tryCatch } from "../common/TryCatchHandler.js";
 import { ValidationError } from "../common/ValidationError.js";
+import { Globals } from "../model/index.js";
 import { MatterCoreSpecificationV1_0 } from "../spec/Specifications.js";
 import { Merge } from "../util/Type.js";
 import { TlvAny } from "./TlvAny.js";
@@ -75,26 +77,35 @@ export class ObjectSchema<F extends TlvFields> extends TlvSchema<TypeFromFields<
         }
     }
 
-    override encodeTlvInternal(writer: TlvWriter, value: TypeFromFields<F>, tag?: TlvTag): void {
+    override encodeTlvInternal(
+        writer: TlvWriter,
+        value: TypeFromFields<F>,
+        tag?: TlvTag,
+        forWriteInteraction?: boolean,
+    ): void {
         writer.writeTag({ type: this.type }, tag);
         for (const name in this.fieldDefinitions) {
             const { id, schema, optional: isOptional, repeated: isRepeated } = this.fieldDefinitions[name];
             const fieldValue = (value as any)[name];
             if (fieldValue === undefined) {
                 if (!isOptional) {
-                    throw new ValidationError(`Missing mandatory field ${name}`);
+                    if (forWriteInteraction && id === <number>Globals.FabricIndex.id) {
+                        // FabricIndex field should not be included in encoded data for write interactions
+                        continue;
+                    }
+                    throw new ValidationError(`Missing mandatory field ${name}`, name);
                 }
                 continue;
             }
             if (isRepeated) {
                 if (!Array.isArray(fieldValue)) {
-                    throw new ValidationError(`Repeated field ${name} should be an array.`);
+                    throw new ValidationError(`Repeated field ${name} should be an array.`, name);
                 }
                 for (const element of fieldValue) {
-                    schema.encodeTlvInternal(writer, element, { id });
+                    schema.encodeTlvInternal(writer, element, { id }, forWriteInteraction);
                 }
             } else {
-                schema.encodeTlvInternal(writer, fieldValue, { id });
+                schema.encodeTlvInternal(writer, fieldValue, { id }, forWriteInteraction);
             }
         }
         writer.writeTag({ type: TlvType.EndOfContainer });
@@ -153,26 +164,42 @@ export class ObjectSchema<F extends TlvFields> extends TlvSchema<TypeFromFields<
                 if (optional) {
                     continue;
                 }
-                throw new ValidationError(`Missing mandatory field ${name}`);
+                throw new ValidationError(`Missing mandatory field ${name}`, name);
             }
             if (isRepeated) {
                 const { minLength = 2, maxLength = 65535 } = this.fieldDefinitions[name] as RepeatedFieldType<any>;
                 if (!Array.isArray(data)) {
-                    throw new ValidationError(`Repeated field ${name} should be an array.`);
+                    throw new ValidationError(`Repeated field ${name} should be an array.`, name);
                 }
                 if (data.length > maxLength)
                     throw new ValidationError(
                         `Repeated field list for ${name} is too long: ${data.length}, max ${maxLength}.`,
+                        name,
                     );
                 if (data.length < minLength)
                     throw new ValidationError(
                         `Repeated field list for ${name} is too short: ${data.length}, min ${minLength}.`,
+                        name,
                     );
                 for (const element of data) {
-                    schema.validate(element);
+                    tryCatch(
+                        () => schema.validate(element),
+                        ValidationError,
+                        error => {
+                            error.fieldName = `${name}${error.fieldName !== undefined ? `.${error.fieldName}` : ""}`;
+                            throw error;
+                        },
+                    );
                 }
             } else {
-                schema.validate(data);
+                tryCatch(
+                    () => schema.validate(data),
+                    ValidationError,
+                    error => {
+                        error.fieldName = `${name}${error.fieldName !== undefined ? `.${error.fieldName}` : ""}`;
+                        throw error;
+                    },
+                );
             }
         }
     }
@@ -207,7 +234,7 @@ export class ObjectSchema<F extends TlvFields> extends TlvSchema<TypeFromFields<
             const field = this.fieldDefinitions[k] as FieldType<any>;
 
             if (field.id === fieldId) {
-                if (removeChecker((value as any)[k])) {
+                if ((value as any)[k] !== undefined && removeChecker((value as any)[k])) {
                     delete (value as any)[k];
                 }
             } else {