+++ packages/matter.js/test/tlv/TlvObjectTest.ts
@@ -4,6 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import { ValidationError } from "../../src/common/ValidationError.js";
+import { FabricIndex, TlvFabricIndex } from "../../src/datatype/FabricIndex.js";
 import { TlvAny } from "../../src/tlv/TlvAny.js";
 import { TlvArray } from "../../src/tlv/TlvArray.js";
 import { TlvNullable } from "../../src/tlv/TlvNullable.js";
@@ -65,6 +67,16 @@ const schemaListRepeatedUnlimited = TlvTaggedList({
     repeatedField: TlvRepeatedField(2, TlvString),
 });
 
+const schemaWithFabricIndex = TlvObject({
+    /** Mandatory field jsdoc */
+    mandatoryField: TlvField(1, TlvUInt8),
+
+    /** Optional field jsdoc */
+    optionalField: TlvOptionalField(2, TlvString),
+
+    fabricIndex: TlvField(254, TlvFabricIndex),
+});
+
 type CodecVector<I, E> = { [valueDescription: string]: { encoded: E; decoded: I } };
 
 const codecVector: CodecVector<TypeFromSchema<typeof schema>, string> = {
@@ -120,6 +132,44 @@ describe("TlvObject tests", () => {
                     expect(tlvDecoded).deep.equal(decoded);
                 });
             }
+        });
+
+        describe("encodeTlv with decodeTlv with fabric index", () => {
+            it(`encode/decodes with fabricIndex`, () => {
+                const tlvEncoded = schemaWithFabricIndex.encodeTlv({
+                    mandatoryField: 1,
+                    optionalField: "test",
+                    fabricIndex: FabricIndex(1),
+                });
+                const tlvDecoded = schemaWithFabricIndex.decodeTlv(tlvEncoded);
+                expect(tlvDecoded).deep.equal({
+                    mandatoryField: 1,
+                    optionalField: "test",
+                    fabricIndex: FabricIndex(1),
+                });
+            });
+
+            it(`encode/decodes with ignoring fabricIndex for write interaction`, () => {
+                const noFabricEncoded = schema.encodeTlv({ mandatoryField: 1, optionalField: "test" });
+
+                const tlvEncoded = schemaWithFabricIndex.encodeTlv(
+                    // @ts-expect-error fabricIndex missing because undefined and ok for writeInteraction
+                    {
+                        mandatoryField: 1,
+                        optionalField: "test",
+                    },
+                    true,
+                );
+
+                // THe Tlv encoded data is the same as without FabricIndex
+                expect(tlvEncoded).deep.equal(noFabricEncoded);
+
+                const tlvDecoded = schemaWithFabricIndex.decodeTlv(tlvEncoded);
+                expect(tlvDecoded).deep.equal({
+                    mandatoryField: 1,
+                    optionalField: "test",
+                });
+            });
         });
 
         describe("inject Field value", () => {
@@ -182,6 +232,24 @@ describe("TlvObject tests", () => {
                     { mandatoryField: 1, optionalField: "test" },
                     { mandatoryField: 2, optionalField: "test" },
                 ]);
+            });
+        });
+
+        describe("ValidationError", () => {
+            it("throws error on invalid field value", () => {
+                let hasThrown = false;
+                try {
+                    // @ts-expect-error  test case
+                    schema.validate({ mandatoryField: 1, optionalField: 2 });
+                } catch (error) {
+                    hasThrown = true;
+                    expect(error instanceof ValidationError).true;
+                    if (error instanceof ValidationError) {
+                        expect(error.message).equal("(Validation/135) Expected string, got number.");
+                        expect(error.fieldName).equal("optionalField");
+                    }
+                }
+                expect(hasThrown).true;
             });
         });